@@ -1,65 +1,67 @@
 {
-  "name": "gatsby-prismic-starter-prist",
+  "name": "shawnohn.com",
   "private": true,
-  "description": "A starter powered by Gatsby and Prismic to showcase portfolios and blogs.",
-  "version": "0.1.0",
-  "author": "Marguerite Roth | marguerite.io | marguerite.roth@gmail.com",
+  "description": "My personal website running on Gatsby",
+  "version": "1.0.0",
+  "author": "Shawn ohn <developer.ohn@gmail.com>",
   "dependencies": {
-    "@emotion/core": "^10.0.10",
-    "@emotion/styled": "^10.0.12",
-    "gatsby": "^2.12.0",
-    "gatsby-image": "^2.2.20",
-    "gatsby-plugin-catch-links": "^2.0.12",
-    "gatsby-plugin-emotion": "^4.1.0",
-    "gatsby-plugin-google-analytics": "^2.1.6",
-    "gatsby-plugin-manifest": "^2.2.4",
-    "gatsby-plugin-offline": "^3.0.8",
-    "gatsby-plugin-react-helmet": "^3.1.8",
-    "gatsby-plugin-resolve-src": "^2.0.0",
-    "gatsby-plugin-sass": "^2.0.11",
-    "gatsby-plugin-sharp": "^2.2.25",
-    "gatsby-source-filesystem": "^2.1.26",
-    "gatsby-source-prismic-graphql": "3.4.0-beta.1",
-    "gatsby-transformer-sharp": "^2.2.16",
-    "node-sass": "^4.12.0",
-    "prismic-reactjs": "^1.1.1",
+    "@mdx-js/mdx": "^1.6.6",
+    "@mdx-js/react": "^1.6.6",
+    "@tailwindcss/typography": "^0.1.3",
+    "gatsby": "^2.21.37",
+    "gatsby-image": "^2.4.4",
+    "gatsby-plugin-env-variables": "^1.0.2",
+    "gatsby-plugin-google-analytics": "^2.3.4",
+    "gatsby-plugin-manifest": "^2.4.5",
+    "gatsby-plugin-mdx": "^1.2.18",
+    "gatsby-plugin-offline": "^3.2.16",
+    "gatsby-plugin-postcss": "^2.3.2",
+    "gatsby-plugin-react-helmet": "^3.3.1",
+    "gatsby-plugin-robots-txt": "^1.5.1",
+    "gatsby-plugin-sharp": "^2.6.4",
+    "gatsby-plugin-sitemap": "^2.4.5",
+    "gatsby-remark-autolink-headers": "^2.3.11",
+    "gatsby-remark-images": "^3.3.13",
+    "gatsby-remark-prismjs": "^3.5.10",
+    "gatsby-source-filesystem": "^2.3.4",
+    "gatsby-transformer-sharp": "^2.5.2",
+    "prismjs": "^1.20.0",
     "prop-types": "^15.7.2",
-    "react": "^16.9.0",
-    "react-anchor-link-smooth-scroll": "^1.0.12",
-    "react-dom": "^16.9.0",
-    "react-emoji": "^0.5.0",
-    "react-helmet": "^5.2.1",
-    "react-markdown": "^4.0.6",
-    "react-modal": "^3.8.1",
-    "react-moment": "^0.9.2",
-    "react-router-dom": "^5.0.0",
-    "react-router-hash-link": "^1.2.1",
-    "smooth-scroll": "^16.0.3"
+    "react": "^16.12.0",
+    "react-dom": "^16.12.0",
+    "react-helmet": "^6.0.0"
   },
   "devDependencies": {
-    "gatsby-plugin-remove-trailing-slashes": "^2.0.11",
-    "prettier": "^1.18.2"
+    "@babel/core": "^7.10.2",
+    "@storybook/addon-a11y": "^5.3.19",
+    "@storybook/addon-actions": "^5.3.19",
+    "@storybook/addon-knobs": "^5.3.19",
+    "@storybook/addon-links": "^5.3.19",
+    "@storybook/addons": "^5.3.19",
+    "@storybook/react": "^5.3.19",
+    "babel-loader": "^8.1.0",
+    "cross-env": "^7.0.2",
+    "prettier": "2.0.5",
+    "storybook-dark-mode": "^0.6.1",
+    "tailwindcss": "^1.6.0"
   },
   "keywords": [
-    "gatsby",
-    "prismic",
-    "blog",
-    "portfolio"
+    "gatsby"
   ],
   "license": "MIT",
   "scripts": {
     "build": "gatsby build",
-    "develop": "gatsby develop",
-    "format": "prettier --write src/**/*.{js,jsx}",
+    "develop": "cross-env GATSBY_GRAPHQL_IDE=playground gatsby develop",
+    "format": "prettier --write \"**/*.{js,jsx,json,md}\"",
     "start": "npm run develop",
     "serve": "gatsby serve",
-    "test": "echo \"Write tests! -> https://gatsby.dev/unit-testing\""
+    "clean": "gatsby clean",
+    "test": "echo \"Write tests! -> https://gatsby.dev/unit-testing\" && exit 1",
+    "storybook": "start-storybook -p 6006",
+    "build-storybook": "build-storybook"
   },
   "repository": {
     "type": "git",
-    "url": "https://github.com/margueriteroth/gatsby-prismic-starter-prist"
-  },
-  "bugs": {
-    "url": "https://github.com/gatsbyjs/gatsby/issues"
+    "url": "https://github.com/shawnohn/shawnohn.com.git"
   }
 }
