@@ -1,10 +1,2 @@
-/**
- * Implement Gatsby's Browser APIs in this file.
- *
- * See: https://www.gatsbyjs.org/docs/browser-apis/
- */
-
-// const { registerLinkResolver } = require('gatsby-source-prismic-graphql');
-// const { linkResolver } = require('./src/utils/prismic-configuration');
-
-// registerLinkResolver(linkResolver);
+import './src/assets/css/index.css'
+require('./src/assets/css/code.css')
