@@ -1,134 +1,23 @@
-import React from "react";
-import PropTypes from "prop-types";
-import Helmet from "react-helmet";
-import { graphql } from "gatsby";
-import styled from "@emotion/styled";
-import dimensions from "styles/dimensions";
-import Layout from "components/Layout";
-import PostCard from "components/PostCard";
+import React from 'react'
+import Layout from '../components/layout'
+import SEO from '../components/seo'
+import PostList from '../components/postList'
+import { Heading } from '../components/atoms'
+import Newsletter from '../components/newsletter'
 
-const BlogTitle = styled("h1")`
-    margin-bottom: 1em;
-`
-
-const BlogGrid = styled("div")`
-    display: grid;
-    grid-template-columns: repeat(3, 1fr);
-    grid-gap: 2.5em;
-
-    @media(max-width: 1050px) {
-        grid-template-columns: repeat(2, 1fr);
-        grid-gap: 1.5em;
-    }
-
-    @media(max-width: ${dimensions.maxwidthMobile}px) {
-        grid-template-columns: 1fr;
-        grid-gap: 2.5em;
-    }
-`
-
-const Blog = ({ posts, meta }) => (
-    <>
-        <Helmet
-            title={`Blog | Prist, Gatsby & Prismic Starter`}
-            titleTemplate={`%s | Blog | Prist, Gatsby & Prismic Starter`}
-            meta={[
-                {
-                    name: `description`,
-                    content: meta.description,
-                },
-                {
-                    property: `og:title`,
-                    content: `Blog | Prist, Gatsby & Prismic Starter`,
-                },
-                {
-                    property: `og:description`,
-                    content: meta.description,
-                },
-                {
-                    property: `og:type`,
-                    content: `website`,
-                },
-                {
-                    name: `twitter:card`,
-                    content: `summary`,
-                },
-                {
-                    name: `twitter:creator`,
-                    content: meta.author,
-                },
-                {
-                    name: `twitter:title`,
-                    content: meta.title,
-                },
-                {
-                    name: `twitter:description`,
-                    content: meta.description,
-                },
-            ].concat(meta)}
-        />
-        <Layout>
-            <BlogTitle>
-                Blog
-            </BlogTitle>
-            <BlogGrid>
-                {posts.map((post, i) => (
-                    <PostCard
-                        key={i}
-                        author={post.node.post_author}
-                        category={post.node.post_category}
-                        title={post.node.post_title}
-                        date={post.node.post_date}
-                        description={post.node.post_preview_description}
-                        uid={post.node._meta.uid}
-                    />
-                ))}
-            </BlogGrid>
-        </Layout>
-    </>
-);
-
-export default ({ data }) => {
-    const posts = data.prismic.allPosts.edges;
-    const meta = data.site.siteMetadata;
-    if (!posts) return null;
-
-    return (
-        <Blog posts={posts} meta={meta}/>
-    )
+const Blog = () => {
+  return (
+    <Layout activePage="blog">
+      <SEO title="Blog" />
+      <Heading>Blog</Heading>
+      <p className="-mt-3 mb-12 text-tertiary">
+        I write about things I learn in web development and also stuff that I
+        find interesting.
+      </p>
+      <PostList />
+      <Newsletter />
+    </Layout>
+  )
 }
 
-Blog.propTypes = {
-    posts: PropTypes.array.isRequired,
-    meta: PropTypes.object.isRequired,
-};
-
-
-export const query = graphql`
-    {
-        prismic {
-            allPosts(sortBy: post_date_DESC) {
-                edges {
-                    node {
-                        post_title
-                        post_date
-                        post_category
-                        post_preview_description
-                        post_author
-                        _meta {
-                            uid
-                        }
-                    }
-                }
-            }
-        }
-        site {
-            siteMetadata {
-                title
-                description
-                author
-            }
-        }
-    }
-`
-
+export default Blog
